NASA_Challenge_[@xfiderek] Add .gitignore to ROS 2 Trick demo (#42).
As part of the NASA Spark Space ROS challenge of 2024, a new demo showing how
Space ROS can be connected to Trick is being developed. That demo will be in a
separate directory, ros_trick, which will hold all files related to the demo.

This commit adds a top-level .gitignore for the Trick demo so that files
generated during development and compilation are not added to the repository by
mistake.