NASA_Challenge_[@xfiderek] Add trick_ros2_control package (#42).
As part of the NASA Spark Space ROS challenge of 2024, a new demo showing how
Space ROS can be connected to Trick is being developed.

This commit adds a ROS 2 package with ROS 2 Control Hardware Plugin that allows
to use Trick as Hardware System. It works similarly to gazebo_ros2_control.