NASA_Challenge_[@xfiderek] Add canadarm_wrench_publisher package (#42).
As part of the NASA Spark Space ROS challenge of 2024, a new demo showing how
Space ROS can be connected to Trick is being developed.

This commit adds a canadarm_wrench_publisher package that demonstrates how to
use the Trick bridge in any ROS 2 node. It is also used in the end-to-end demo.