Pin versions of upstream deps used in dockerfile cloned from spaceros repo (#42).
The ROS 2 - Trick demo was prepared to work with an older version of Space ROS,
but the project has since moved to Jazzy.

This commit adds a build script that pins the version of the dockerfile to a
specific version of Space ROS that the demo is known to work with, and modifies
the build script to download and patch dependencies as needed.